fix(build): Suppress irrelevant JSC_UNUSED_LOCAL_ASSIGNMENT errors
tsc generates code for merged namespaces that looks like:

    (function (ClassName) {
        let EnumName;
        (function (EnumName) {
            EnumName[EnumNameAlign["v1"] = 0] = "v1";
            // etc.
        })(EnumName = ClassName.EnumName || (ClassName.EnumName = {}));
    })(ClassName || (ClassName = {}));

and Closure Compiler complains about the fact that the EnumName let
binding is initialised but never used.  (It exists so that any other
code that was in the namespace could see the enum.)

Suppress this message, since it is not actionable and lint and/or tsc
should tell us if we have actual unused variables in our .ts files.