feat(build): Have npm scripts run npm ci first where applicable
Have any npm script that have external effects (e.g. publishing an
npm package, pushing a new version to appengine, or updating GitHub
Pages) start by running npm ci to ensure that all dependencies are
up-to-date with respect to package-lock.json.

(This is done by npm and not a gulp script because gulp itself
might need updating.  So might npm, but that is less likely to
make any difference to what gets published/pushed.)