chore(tests): Use freshly-build files in compressed mode.
Use the freshly-built build/*_compresssed.js files when bootstrapping
in compressed mode, rather than using the checked-in files in the
repository root.

This helps ensure that compressed and uncompressed mode will be
testing (as closely as possible) the same code.

Obsoletes #6218 (though the issues discussed there have not actually
yet been addressed in this branch).