fix(tests): Have npm start run a full build
Doing npm start in a clean checkout loads a broken playground,
because the deps script does not run buildLangfiles.

To fix this (and also to ensure that compressed playground loading
also works, even though uncompressed is the default), have the
start script run the build script rather than just deps.