chore(tests): Remove circular import loading issue kludge
Prior to PR #6818, circular imports resulted in the debug module
loader (in closure/goog/base.js) failing to record the
goog.module ID of most modules that were
involved in the cycle, and in particular of the Blockly.Xml
module.  This had secondary fallout that resulted
in library blocks modules being loaded in the wrong order.

A kludge was introduced in PR #6703 that worked around this
problem by making sure that window.Blockly was set, allowing
the modules loaded out-of-order to still work.

Now that we have removed all remaining circular dependencies
there is no need for the kludge, since all module IDs are
properly recorded and modules are loaded in the correct order.