refactor(generators)!: Update generator definitions to use dictionary
BREAKING CHANGE: this PR moves the generator functions we provide
from their previous location directly on the CodeGenerator instances
to the new .forBlock dictionary on each instance.

The change in the previous commit does not oblige external
developers to do the same for their custom generators, but
they will need to update any code that references the generator
functions we provide in generators/*/* (e.g. to reuse one of
our functions for a different block type).