fix(generators): Fix more minor inconsistencies in JS and Python
The migration of the JavaScript and Python generators
inadvertently introduced some inconsistencies in the code,
e.g. putting executable code before the initial comment line that
most generator functions begin with.  This fixes another instance
of this (but n.b. that these inline comments should have been
JSDocs and a task has been added to #7600 to convert them).

Additionally, I noticed while doing the PHP migration that
ORDER_OVERRIDES was not typed as specifically as it could be,
in previous migrations, so this is fixed here (along with the
formatting of the associated JSDoc, which can fit on one line
now.)