fix(typings): Remove bogus .d.ts files; add new languages
PR #3821 added .d.ts files for every file in msg/json/, but several
of these are internal utility files rather than translations, and
do not result in a langfile being output by create_messages.py
when building langfiles.

In the meantime we have added a few new languages that are being
published but which have (until now) not had the corresponding
type declarations.