fix(typings): Restore inadvertently-deleted typings/msg/msg.d.ts
Fixes #7952.

The main typings file for langfiles, typings/msg/msg.d.ts, was
inadvertently deleted in PR #7822.

This was part of a well intentioned attempt to remove spurious
files from typings/msg/ that do not correspond to published
langfiles, but this file should have been retained because
msg.d.ts is reexported by all the other *.d.ts files in this
directory.