feat!: initial integration of FocusableTree/Node.
Much more work is needed, and this is just a staged example of where
things are moving. The actual introduction of FocusableTree and
FocusableNode implementations will happen in isolation, but this is
exploring other changes that may be needed in order to bring
FocusManager to the keyboard nav plugin with current functional parity.